feat: add fractional diameters (½ ¾ 1½ 1¾ 2½ 3½) to DN_UNION enum
Backend accepts float dn_union (was int-only, blocked fractions with 422).
Frontend renders Unicode fractions with aria-labels for screen readers.
parseInt → parseFloat in DN handlers to preserve decimal values.

+++ backend/routers/union_router.py
@@ -569,7 +569,7 @@ async def guardar_uniones(
                 to_create.append({"n_union": n, "dn_union": u.dn_union, "tipo_union": u.tipo_union})
             else:
                 ex = existing_by_n[n]
-                ex_dn = int(ex["dn_union"]) if ex["dn_union"] is not None else None
+                ex_dn = float(ex["dn_union"]) if ex["dn_union"] is not None else None
                 ex_tipo = ex["tipo_union"]
                 changed = ex_dn != u.dn_union or ex_tipo != u.tipo_union
                 # HIGH-3: Block dn_union/tipo_union changes on unions with work
